pop up done
moved grid filler to its class

pop up form is done
TODO: pop up the status code's context from enum class instead of code itself

implement rejected mail thing
implement ok status thing

+++ Items/Log/LogEntry.cs
@@ -9,7 +9,7 @@
         public string? username { get; set; }
         public string? accessLocation { get; set; }
         public int accessDirection { get; set; }
-        public int verifyStatusCode { get; set; }
+        public UInt16 verifyStatusCode { get; set; }
         public string? additionalInfo { get; set; }
         public string? logTime { get; set; }
     }
